Fix a bunch of compiler warnings
git-svn-id: https://svn.code.sf.net/p/plexil/code/branches/plexil-4@4906 d2be67b0-814c-0410-949f-d215b6b29573

+++ src/utils/test/LinkedQueueTest.cc
@@ -155,7 +155,7 @@ static bool testLinkedQueueBasics()
   for (int i = 1; i <= n; ++i) {
     testq.push(new QueueTest(i));
     assertTrue_1(!testq.empty());
-    assertTrue_1(testq.size() == i);
+    assertTrue_1(testq.size() == (size_t) i);
     assertTrue_1(testq.front() != NULL);
   }
 
@@ -170,7 +170,7 @@ static bool testLinkedQueueBasics()
   for (int i = 1; i <= n; ++i) {
     testq.push(new QueueTest(i));
     assertTrue_1(!testq.empty());
-    assertTrue_1(testq.size() == i);
+    assertTrue_1(testq.size() == (size_t) i);
     assertTrue_1(testq.front() != NULL);
   }
 
@@ -192,7 +192,7 @@ static bool testLinkedQueueBasics()
   assertTrue_1(nxt != NULL);
   testq.remove(item);
   assertTrue_1(!testq.empty());
-  assertTrue_1(testq.size() == n - 1);
+  assertTrue_1(testq.size() == (size_t) (n - 1));
   assertTrue_1(testq.front() == nxt);
   delete item;
 
@@ -201,7 +201,7 @@ static bool testLinkedQueueBasics()
   assertTrue_1(item != NULL);
   testq.remove(item);
   assertTrue_1(!testq.empty());
-  assertTrue_1(testq.size() == n - 2);
+  assertTrue_1(testq.size() == (size_t) (n - 2));
   delete item;
 
   // Remove from end
@@ -211,14 +211,14 @@ static bool testLinkedQueueBasics()
   // item now points to last
   testq.remove(item);
   assertTrue_1(!testq.empty());
-  assertTrue_1(testq.size() == n - 3);
+  assertTrue_1(testq.size() == (size_t) (n - 3));
   delete item;
 
   // Try to "remove" an item not in queue
   item = new QueueTest(42);
   testq.remove(item);
   assertTrue_1(!testq.empty());
-  assertTrue_1(testq.size() == n - 3);
+  assertTrue_1(testq.size() == (size_t) (n - 3));
   delete item;
 
   // Pop and delete remaining
@@ -281,7 +281,7 @@ static bool testLinkedQueueFindIf()
     testq.push(new QueueTest(i));
     assertTrue_1(!testq.empty());
     assertTrue_1(testq.front() != NULL);
-    assertTrue_1(testq.size() == i);
+    assertTrue_1(testq.size() == (size_t) i);
   }
 
   // Try finding first item
@@ -331,7 +331,7 @@ static bool testLinkedQueueRemoveIf()
     testq.push(new QueueTest(i));
     assertTrue_1(!testq.empty());
     assertTrue_1(testq.front() != NULL);
-    assertTrue_1(testq.size() == i);
+    assertTrue_1(testq.size() == (size_t) i);
   }
 
   // Try removing from front
@@ -343,7 +343,7 @@ static bool testLinkedQueueRemoveIf()
   assertTrue_1(item->value == 1);
   assertTrue_1(item == temp);
   assertTrue_1(!testq.empty());
-  assertTrue_1(testq.size() == n - 1);
+  assertTrue_1(testq.size() == (size_t) (n - 1));
   assertTrue_1(testq.front()->value == 2);
   delete item;
 
@@ -355,7 +355,7 @@ static bool testLinkedQueueRemoveIf()
   assertTrue_1(item->value == 10);
   assertTrue_1(item == temp);
   assertTrue_1(!testq.empty());
-  assertTrue_1(testq.size() == n - 2);
+  assertTrue_1(testq.size() == (size_t) (n - 2));
   assertTrue_1(testq.front()->value == 2);
   delete item;
 
@@ -367,7 +367,7 @@ static bool testLinkedQueueRemoveIf()
   assertTrue_1(item->value == 6);
   assertTrue_1(item == temp);
   assertTrue_1(!testq.empty());
-  assertTrue_1(testq.size() == n - 3);
+  assertTrue_1(testq.size() == (size_t) (n - 3));
   assertTrue_1(testq.front()->value == 2);
   delete item;
 
@@ -429,7 +429,7 @@ static bool testPriorityQueue()
     // Test insert on empty queue
     testpq.insert(won);
     assertTrue_1(!testpq.empty());
-    assertTrue_1(testpq.size() == 1);
+    assertTrue_1(testpq.size() == (size_t) 1);
     assertTrue_1(testpq.front() != NULL);
     assertTrue_1(testpq.front()->value == 1);
     assertTrue_1(testpq.front()->next() == NULL);
@@ -443,7 +443,7 @@ static bool testPriorityQueue()
     // Insert again
     testpq.insert(won);
     assertTrue_1(!testpq.empty());
-    assertTrue_1(testpq.size() == 1);
+    assertTrue_1(testpq.size() == (size_t) 1);
     assertTrue_1(testpq.front() != NULL);
     assertTrue_1(testpq.front()->value == 1);
     assertTrue_1(testpq.front()->next() == NULL);
@@ -451,7 +451,7 @@ static bool testPriorityQueue()
     // Remove of NULL should be safe
     testpq.remove(NULL);
     assertTrue_1(!testpq.empty());
-    assertTrue_1(testpq.size() == 1);
+    assertTrue_1(testpq.size() == (size_t) 1);
     assertTrue_1(testpq.front()->value == 1);
 
     // Test removal of only item
@@ -474,7 +474,7 @@ static bool testPriorityQueue()
   for (int i = 1; i <= n; ++i) {
     testpq.insert(new QueueTest(i));
     assertTrue_1(!testpq.empty());
-    assertTrue_1(testpq.size() == i);
+    assertTrue_1(testpq.size() == (size_t) i);
     assertTrue_1(testpq.front() != NULL);
   }
 
@@ -496,7 +496,7 @@ static bool testPriorityQueue()
   for (int i = 1; i <= n; ++i) {
     testpq.insert(new QueueTest(n + 1 - i));
     assertTrue_1(!testpq.empty());
-    assertTrue_1(testpq.size() == i);
+    assertTrue_1(testpq.size() == (size_t) i);
     assertTrue_1(testpq.front() != NULL);
   }
 
@@ -519,7 +519,7 @@ static bool testPriorityQueue()
     testpq.insert(new QueueTest(n + 1 - i));
     testpq.insert(new QueueTest(n/2 + 1 - i));
     assertTrue_1(!testpq.empty());
-    assertTrue_1(testpq.size() == 2 * i);
+    assertTrue_1(testpq.size() == (size_t) (2 * i));
     assertTrue_1(testpq.front() != NULL);
   }
 
@@ -542,13 +542,13 @@ static bool testPriorityQueue()
   assertTrue_1(temp != NULL);
   testpq.remove(item);
   assertTrue_1(!testpq.empty());
-  assertTrue_1(testpq.size() == n - 1);
+  assertTrue_1(testpq.size() == (size_t) (n - 1));
   assertTrue_1(testpq.front() == temp);
 
   // Reinsert and check that it winds up in front
   testpq.insert(item);
   assertTrue_1(!testpq.empty());
-  assertTrue_1(testpq.size() == n);
+  assertTrue_1(testpq.size() == (size_t) n);
   assertTrue_1(item == testpq.front());
 
   // Remove from middle
@@ -556,12 +556,12 @@ static bool testPriorityQueue()
   assertTrue_1(item != NULL);
   testpq.remove(item);
   assertTrue_1(!testpq.empty());
-  assertTrue_1(testpq.size() == n - 1);
+  assertTrue_1(testpq.size() == (size_t) (n - 1));
 
   // Reinsert and check that it winds up where it was
   testpq.insert(item);
   assertTrue_1(!testpq.empty());
-  assertTrue_1(testpq.size() == n);
+  assertTrue_1(testpq.size() == (size_t) n);
   assertTrue_1(item == testpq.front()->next()->next()->next());
 
   // Remove from end
@@ -569,12 +569,12 @@ static bool testPriorityQueue()
   assertTrue_1(item);
   testpq.remove(item);
   assertTrue_1(!testpq.empty());
-  assertTrue_1(testpq.size() == n - 1);
+  assertTrue_1(testpq.size() == (size_t) (n - 1));
 
   // Insert it back and check it winds up at the end
   testpq.insert(item);
   assertTrue_1(!testpq.empty());
-  assertTrue_1(testpq.size() == n);
+  assertTrue_1(testpq.size() == (size_t) n);
   temp = testpq.find_if(EqualsNull());
   assertTrue_1(temp == item);
 
@@ -582,7 +582,7 @@ static bool testPriorityQueue()
   item = new QueueTest(42);
   testpq.remove(item);
   assertTrue_1(!testpq.empty());
-  assertTrue_1(testpq.size() == n);
+  assertTrue_1(testpq.size() == (size_t) n);
   delete item;
 
   // Insert duplicate items and check they wind up behind originals
@@ -590,16 +590,16 @@ static bool testPriorityQueue()
   temp = new QueueTest(*item);
   testpq.insert(temp);
   assertTrue_1(!testpq.empty());
-  assertTrue_1(testpq.size() == n + 1);
+  assertTrue_1(testpq.size() == (size_t) (n + 1));
   assertTrue_1(item->next() == temp);
 
   testpq.remove(temp);
-  assertTrue_1(testpq.size() == n);
+  assertTrue_1(testpq.size() == (size_t) n);
   item = testpq.find_if(EqualsSix());
   assertTrue_1(item->value == 6);
   temp->value = 6;
   testpq.insert(temp);
-  assertTrue_1(testpq.size() == n + 1);
+  assertTrue_1(testpq.size() == (size_t) (n + 1));
   assertTrue_1(item->next() == temp);
 
   temp = NULL;
